thorgodoflightning/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
python-dotenv>=1.0.0
firebase-admin>=6.2.0
google-cloud-storage>=2.10.0
google-cloud-videointelligence>=2.11.0
google-cloud-vision>=3.4.0
openai>=1.0.0
python-multipart>=0.0.6
pydantic>=2.4.2
pinecone-client>=2.2.4
boto3>=1.28.0
paapi5-python-sdk>=1.1.0 